# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Gudang API",
        "description": "Dokumentasi API untuk aplikasi sistem gudang berbasis Laravel.",
        "contact": {
            "email": "munirrmiftahul94@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://sistem-gudang.test",
            "description": "Local Development"
        },
        {
            "url": "https://api.sistem-gudang.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/pegawai-action/all": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Ambil daftar pegawai",
                "operationId": "0d1528ef9d25b343cb382842847f8252",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar employee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "name": "string"
                                        },
                                        {
                                            "name": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pegawai": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Employee",
                "description": "Employee",
                "operationId": "5adc326f00276bbd6bfd5b6640431d76",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "description": "credentials",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "pegawai_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Hartono"
                                    },
                                    "pegawai_email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "hartono@gmail.com"
                                    },
                                    "pegawai_phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "085"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar lokasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "name": "string",
                                            "email": "string",
                                            "phone": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pegawai/{id}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Lihat detail pegawai",
                "description": "Mengambil data pegawai berdasarkan ID",
                "operationId": "b76e696b13ad582b910556173f379f64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pegawai",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "6cecf10af2584e8ea7c46fdaab339978"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pegawai ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "pegawai_name": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lokasi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Lokasi tidak ditemukan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Anda. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}